CORS-4281: Remove the GCP Service Endpoints
Removing the GCP Service Endpoints in favor of a new solution. The new solution will create a
private hosted zone that will route traffic to the googleapi endpoints via an ip address created
during the initialization of the private serivce connect endpoint. The cluster components no longer
need the endpoint overrides, so the service endpoints can and should be removed to avoid confusion.